requirements.txt
groq==1.1.2
python-dotenv